refactor: extract circuit/ subpackage (_circuit.py + einsum_cache.py)
- src/pdft/_circuit.py     -> src/pdft/circuit/builder.py
- src/pdft/einsum_cache.py -> src/pdft/circuit/cache.py
- tests/test_einsum_cache.py -> tests/circuit/test_cache.py

All basis files updated to import from .circuit.builder / .circuit.cache.
Pure file moves + import rewrites; no logic change.

Refs: docs/superpowers/specs/2026-04-26-src-modular-refactor-design.md
